frame-a/pyproject.toml
[project]
authors = [
    {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
description = ""
name = ""
requires-python = ">=3.12"
version = ""
dependencies = [
    "strawberry-graphql>=0.213.0",
    "fastapi",
    "uvicorn",
]
readme = "README.md"
license = {text = "MIT"}

[tool.black]
line-length = 88

[tool.isort]
combine_as_imports = true
combine_star = true
indent = 4
known_pytest = ["pytest"]
known_strawberry = ["strawberry"]
lines_after_imports = 2
profile = "black"
sections = ["FUTURE", "STDLIB", "PYTEST", "THIRDPARTY", "STRAWBERRY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]

[tool.pdm.dev-dependencies]
dev = ["mypy", "pytest", "httpx"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
dev = "uvicorn main:app --reload --port 4001"